[Fix/#45] ScheduleUiModel dDay 타입 Long으로 변경

+++ app/src/main/java/com/swyp/firsttodo/presentation/todo/TodoViewModel.kt
@@ -193,7 +193,7 @@ class TodoViewModel
                         val newSchedules = list.map { model ->
                             ScheduleUiModel(
                                 scheduleId = model.scheduleId,
-                                dDay = model.dDay.toInt(),
+                                dDay = model.dDay,
                                 title = model.title,
                                 date = model.scheduleDate.toDisplayDate(),
                                 rawDate = model.scheduleDate.removeDashes(),
